New features and code updates
Added a text file "bad_words", which contains inappropriate phrases that are banned.
Fixed ChatClient to exit the program upon being disconnected from the server.
Created multiple methods to create a singleton instance of a list containing the inappropriate words in ChatServer
Updated ChatServer code - other classes in the same package can now run the stopServerProcesses() method as long as they have an instance of the server.
Few bugfixes in ChatServer.
Informational messages are now sent to the user upon joining.
Changed the Client class to be more user-friendly and fixed a few bugs.
Added a warning system to Connection. This now allows users to be warned (by the server manager in the near future).
Changed HandleClientInput significantly.
There is now a static method "serverBroadcast" that allows classes to broadcast to the everyone.
User input is now filtered for inappropriate words. If an inappropriate word is found in the user's message, they're warned and the message is not broadcasted.
New, better way to remove connections from the list of connections.
HandleClientInput now stores the instance of the server that the client is connected to.